Model_4/requirements.txt
unstructured
langchain
faiss-cpu
langchain-google-genai
python-dotenv
